TDQ-11150 No indicators show in the studio

+++ main/plugins/org.talend.dataprofiler.core/src/org/talend/dataprofiler/core/manager/DQStructureManager.java
@@ -381,10 +381,6 @@ public final class DQStructureManager {
                             sourcefolder.getProperty().getItem().getState().setPath(subSourceFolder);
                         }
                     }
-                    if (targetfolder.members().length != 0) {
-                        // already initialized
-                        continue;
-                    }
                     copyFilesToFolder(plugin, currentPath, recurse, sourcefolder, suffix, type);
                     continue;
                 }
@@ -411,17 +407,16 @@ public final class DQStructureManager {
                     continue;
                 }
                 IFolder folder = project.getFolder(folderName);
-                if (folder.members().length == 0) {
-                    if (type.equals(ERepositoryObjectType.TDQ_SOURCE_FILE_ELEMENT)) {
-                        String name = file.getName();
-                        int indexOf = name.indexOf("."); //$NON-NLS-1$
-                        String label = name.substring(0, indexOf);
-                        String extendtion = name.substring(indexOf + 1);
-                        createSourceFileItem(file, Path.EMPTY, label, extendtion);
-                    } else {
-                        copyFileToFolder(openStream, fileName, folder);
-                    }
+                if (type.equals(ERepositoryObjectType.TDQ_SOURCE_FILE_ELEMENT)) {
+                    String name = file.getName();
+                    int indexOf = name.indexOf("."); //$NON-NLS-1$
+                    String label = name.substring(0, indexOf);
+                    String extendtion = name.substring(indexOf + 1);
+                    createSourceFileItem(file, Path.EMPTY, label, extendtion);
+                } else {
+                    copyFileToFolder(openStream, fileName, folder);
                 }
+
                 openStream.close();
             } catch (IOException e) {
                 log.error(e, e);
@@ -453,8 +448,14 @@ public final class DQStructureManager {
             if (modelElement != null) {
                 AElementPersistance writer = ElementWriterFactory.getInstance().create(elementFile.getFileExtension());
                 if (writer != null) {
-                    writer.create(modelElement, folder, isImportItem);
-                    elementFile.delete(true, null);
+                    Property initProperty = writer.initProperty(modelElement);
+                    String propFileName = initProperty.getLabel() + "_" + initProperty.getVersion() + "." //$NON-NLS-1$ //$NON-NLS-2$
+                            + elementFile.getFileExtension();
+                    IFile file = elementFile.getParent().getFile(new Path(propFileName));
+                    if (file == null || !file.exists()) {
+                        writer.create(modelElement, folder, isImportItem);
+                        elementFile.delete(true, null);
+                    }
                 }
             }
         }
@@ -479,12 +480,18 @@ public final class DQStructureManager {
             ExceptionHandler.process(e);
         }
         sourceFileItem.setContent(byteArray);
-        IProxyRepositoryFactory repositoryFactory = ProxyRepositoryFactory.getInstance();
-        try {
-            property.setId(repositoryFactory.getNextId());
-            repositoryFactory.create(sourceFileItem, path);
-        } catch (PersistenceException e) {
-            ExceptionHandler.process(e);
+
+        String sqlFileName = property.getLabel() + "_" + property.getVersion() + "." //$NON-NLS-1$ //$NON-NLS-2$
+                + extension;
+        IFile file = ResourceManager.getSourceFileFolder().getFile(path.append(new Path(sqlFileName)));
+        if (file == null || !file.exists()) {
+            IProxyRepositoryFactory repositoryFactory = ProxyRepositoryFactory.getInstance();
+            try {
+                property.setId(repositoryFactory.getNextId());
+                repositoryFactory.create(sourceFileItem, path);
+            } catch (PersistenceException e) {
+                ExceptionHandler.process(e);
+            }
         }
         return sourceFileItem;
     }
